% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{add_p_vals_to_cmap_para}
\alias{add_p_vals_to_cmap_para}
\title{Same as \code{\link{add_p_vals_to_cmap_single}} but parallel}
\usage{
add_p_vals_to_cmap_para(cmap_res, perm_res)
}
\arguments{
\item{cmap_res}{output results from \code{\link{single_broad_wrapper}}}

\item{perm_res}{output results from \code{\link{get_random_cmap_scores}}}
}
\value{

}
\description{
Same as \code{\link{add_p_vals_to_cmap_single}} but parallel
}
