% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{add_p_vals_to_cmap_single}
\alias{add_p_vals_to_cmap_single}
\title{Calculates p-values for CMap results}
\usage{
add_p_vals_to_cmap_single(cmap_res, perm_res)
}
\arguments{
\item{cmap_res}{output results from \code{\link{single_broad_wrapper}}}

\item{perm_res}{output results from \code{\link{get_random_cmap_scores}}}
}
\value{

}
\description{
This function requires the output CMap results as created by \code{\link{single_broad_wrapper}} and a
matrix of permuation values created by \code{\link{get_random_cmap_scores}}. The output is a data.frame
of the CMap results with the p-values added.
}
