% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{best_score_mat}
\alias{best_score_mat}
\title{Chooses the best score per compound}
\usage{
best_score_mat(cs_mat)
}
\arguments{
\item{cs_mat}{data.frame of CS scores}
}
\value{

}
\description{
Chooses the best score per compound
}
