% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{best_score_top20}
\alias{best_score_top20}
\title{Returns the top 20 compounds using the best score approach}
\usage{
best_score_top20(best_cs_mat)
}
\arguments{
\item{best_cs_mat}{output from \link{best_score_mat}}
}
\value{

}
\description{
Returns the top 20 compounds using the best score approach
}
