% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{broad_cmap_score}
\alias{broad_cmap_score}
\title{Calculates CMAP score according to \href{https://clue.io/connectopedia/cmap_algorithms}{CLUE}}
\usage{
broad_cmap_score(up_genes, down_genes, drug_sig_list, n_cores)
}
\arguments{
\item{up_genes}{vector of up-regulated genes}

\item{down_genes}{vector of down-regulated genes}

\item{drug_sig_list}{list of drug signatures}

\item{n_cores}{number of cores}
}
\value{

}
\description{
Calculates CMAP score according to \href{https://clue.io/connectopedia/cmap_algorithms}{CLUE}
}
