% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{calculate_ncs}
\alias{calculate_ncs}
\title{Normalizes the connectivity score by cell type}
\usage{
calculate_ncs(cs_mat)
}
\arguments{
\item{cs_mat}{data.frame of CS scores}
}
\value{

}
\description{
Scores are normalized by cell type.
}
