% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{calculate_pert_cell_score}
\alias{calculate_pert_cell_score}
\title{Calculates pert score per cell type}
\usage{
calculate_pert_cell_score(cs_mat, quant_thresh = 0.67, quant_type = 5)
}
\arguments{
\item{cs_mat}{data.frame of CS scores}

\item{quant_thresh}{percentile threshold. Default is .67 which was used in Subramanian 2017}

\item{quant_type}{This is how the quantiles are calculated. See \link[stats]{quantile}}
}
\value{

}
\description{
Calculates pert score per cell type
}
