% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{calculate_pert_score}
\alias{calculate_pert_score}
\title{Calculates the perturbagen summary score}
\usage{
calculate_pert_score(pert_cell_mat, quant_thresh = 0.67, quant_type = 5)
}
\arguments{
\item{pert_cell_mat}{output from \link{calculate_pert_cell_score}}

\item{quant_thresh}{percentile threshold. Default is .67 which was used in Subramanian 2017}

\item{quant_type}{This is how the quantiles are calculated. See \link[stats]{quantile}}
}
\value{

}
\description{
Calculates the perturbagen summary score
}
