% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunker}
\alias{chunker}
\title{Splits a vector into chunks}
\usage{
chunker(x, num_chunks = 3)
}
\arguments{
\item{x}{vector}

\item{num_chunks}{number of chunks}
}
\value{
list of chunked vectors
}
\description{
Splits a vector into chunks
}
