% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cmap_fgsea_wrapper}
\alias{cmap_fgsea_wrapper}
\title{Wrapper to run the \code{\link{broad_cmap_score}} for many gene sets}
\usage{
cmap_fgsea_wrapper(query_genes, drug_sig_list, n_cores)
}
\arguments{
\item{query_genes}{data.frame of gene signatures}

\item{drug_sig_list}{list of drug signatures}

\item{n_cores}{number of cores}
}
\value{

}
\description{
Wrapper to run the \code{\link{broad_cmap_score}} for many gene sets
}
