% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_annot_leg_params}
\alias{create_annot_leg_params}
\title{Creates legend paramaters for \code{\link[ComplexHeatmap]{HeatmapAnnotation}}}
\usage{
create_annot_leg_params(
  tit = NULL,
  leg_size = 7.5,
  leg_width = NULL,
  leg_height = NULL,
  font_size = 14,
  direction = "vertical"
)
}
\arguments{
\item{tit}{title}

\item{leg_size}{legend size (in mm)}

\item{leg_width}{legend width (in mm)}

\item{leg_height}{legend height  (in mm)}

\item{font_size}{font size}

\item{direction}{direction of legend (either horizontal or vertical )}
}
\value{

}
\description{
The reason I made a function here is that I could never remember how the
parameter list is supposed to be formatted. So it's somewhat superfluous except
that this is hard to google.
}
