% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_gsea_plot_data}
\alias{create_gsea_plot_data}
\title{Creates a data.frame to make a GSEA plot}
\usage{
create_gsea_plot_data(query_idx, ordered_vec, gsea_weight = 0)
}
\arguments{
\item{query_idx}{indexes of query which correpond to position in \code{\link{ordered_vec}}}

\item{ordered_vec}{ordered vector of scores}

\item{gsea_weight}{power to raise the ordered vector (see \code{\href{https://www.genepattern.org/modules/docs/GSEAPreranked/1}{here}} for details.)}
}
\value{

}
\description{
This function creates a data.frame which can be used to
create the line plot of the enrichment score. It was taken
and slightly modified from the \code{\link[fgsea]{plotEnrichment}}.
}
