% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_heatmap_leg_params}
\alias{create_heatmap_leg_params}
\title{Creates legend paramaters for \code{\link[ComplexHeatmap]{Heatmap}}}
\usage{
create_heatmap_leg_params(
  tit = NULL,
  leg_size = 50,
  font_size = 14,
  direction = "horizontal"
)
}
\arguments{
\item{tit}{title}

\item{leg_size}{legend size (in mm)}

\item{font_size}{font size}

\item{direction}{direction of legend (either horizontal or vertical )}
}
\value{

}
\description{
Same sort of function as \code{\link[QSPpaper]{create_annot_leg_params}}
}
