% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{gctx_to_sigList}
\alias{gctx_to_sigList}
\title{Converts GCTX file to list signature list}
\usage{
gctx_to_sigList(gctx_file, cids, decreasing_order = TRUE)
}
\arguments{
\item{gctx_file}{gctx file}

\item{cids}{signatures to parse (see \code{\link{parse.gctx}})}

\item{decreasing_order}{whether to sort in decreasing order}
}
\value{
signature list formatted for \code{\link[QSPpaper]{broad_cmap_score}}
}
\description{
Converts GCTX file to list signature list
}
