% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{get_random_cmap_scores}
\alias{get_random_cmap_scores}
\title{Runs the permutation testing to get a vector of CMAP scores}
\usage{
get_random_cmap_scores(
  up_genes,
  down_genes,
  gctx_file,
  cids = NULL,
  n_perm = 1000,
  num_chunks = 4,
  seed = NULL,
  n_cores = 3
)
}
\arguments{
\item{up_genes}{vector of up-regulated genes}

\item{down_genes}{vector of down-regulated genes}

\item{gctx_file}{gctx file}

\item{cids}{signatures to parse (see \code{\link{parse.gctx}})}

\item{n_cores}{number of cores}
}
\value{

}
\description{
Runs the permutation testing to get a vector of CMAP scores
}
