% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{kegg_gsea_by_coef}
\alias{kegg_gsea_by_coef}
\title{Wrapper to do GSEA on tidy results}
\usage{
kegg_gsea_by_coef(res_df, nPerm = 10000, pvalueCutoff = 0.1)
}
\arguments{
\item{res_df}{a tidy results data.frame of genes (see \code{\link[basicOmics]{get_limma_results}})}

\item{nPerm}{number of permutations to run}

\item{pvalueCutoff}{p-value cutoff of output to retain}
}
\value{
a tidy results data.frame of KEGG pathways
}
\description{
Wrapper to do GSEA on tidy results
}
