% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{lincs2017_score_top20}
\alias{lincs2017_score_top20}
\title{Returns the top 20 compounds after performing maximum quantile statistic}
\usage{
lincs2017_score_top20(pert_mat)
}
\arguments{
\item{pert_mat}{}
}
\value{

}
\description{
Returns the top 20 compounds after performing maximum quantile statistic
}
