% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{make_gsea_plot}
\alias{make_gsea_plot}
\title{Creates a running sum GSEA plot}
\usage{
make_gsea_plot(plot_data, main_title = "")
}
\arguments{
\item{plot_data}{data.frame created by \code{\link{create_gsea_plot_data}}}

\item{main_title}{optional title to be added to plot}
}
\value{

}
\description{
This is my own customized version of the plot created
by \code{\link[fgsea]{plotEnrichment}}.
}
