% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{make_permute_sort_df}
\alias{make_permute_sort_df}
\title{Internal function to create a data.frame with the ordered permuations}
\usage{
make_permute_sort_df(gctx_file, n_perm = 1000, num_chunks = 4, seed = NULL)
}
\arguments{
\item{n_perm}{number of permutations to perform}

\item{num_chunks}{number of chunk levels for chunk_idx column}

\item{seed}{random number seed}
}
\value{

}
\description{
Internal function to create a data.frame with the ordered permuations
}
