% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{single_broad_wrapper}
\alias{single_broad_wrapper}
\title{Wrapper to run the \code{\link{broad_cmap_score}} for a single gene set}
\usage{
single_broad_wrapper(up_genes, down_genes, drug_sig_list, n_cores)
}
\arguments{
\item{up_genes}{vector of up-regulated genes}

\item{down_genes}{vector of down-regulated genes}

\item{drug_sig_list}{list of drug signatures}

\item{n_cores}{number of cores}
}
\value{

}
\description{
This wrapper gets the CMAP scores for a single up and down-regulated gene set accross  all drug signatures
}
