% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniquefy_res_by_group}
\alias{uniquefy_res_by_group}
\title{Uniquefys a res df by group by taking absolute max}
\usage{
uniquefy_res_by_group(
  res_df,
  group_col = "external_gene_name",
  val_col = "GeneRank"
)
}
\arguments{
\item{res_df}{a tidy results data.frame (see \code{\link[delutils]{get_limma_results}})}

\item{group_col}{column with duplicates}

\item{val_col}{column of unique values}
}
\value{
a uniquefyd results data.frame
}
\description{
Uniquefys a res df by group by taking absolute max
}
