% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{broad_cmap_score}
\alias{broad_cmap_score}
\title{Calculates CMAP score using the Broad's latest methodology for a single instance}
\usage{
broad_cmap_score(up_genes, down_genes, drug_sig_vec)
}
\arguments{
\item{up_genes}{vector of up-regulated genes}

\item{down_genes}{vector of down-regulated genes}

\item{drug_sig_vec}{drug signature vector}
}
\value{

}
\description{
This function calculates the CMAP score according to the method outlined  \code{\href{https://clue.io/connectopedia/cmap_algorithms}{here}} for a single instance of upregulated-genes, down-regulated genes and drug-signature
}
